fixed css in EditAudio and AudioPlayerMounter .jsx -> .tsx
