+++ src/components/SlateEditor/plugins/embed/AudioPlayerMounter.tsx
@@ -1,0 +1,76 @@
+/**
+ * Copyright (c) 2018-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import React, { Fragment, useEffect } from 'react';
+import { injectT } from '@ndla/i18n';
+import { initAudioPlayers } from '@ndla/article-scripts';
+// @ts-ignore
+import { AudioPlayer, FigureCaption, FigureLicenseDialog } from '@ndla/ui';
+import { getLicenseByAbbreviation } from '@ndla/licenses';
+import { Audio, TranslateType } from '../../../../interfaces';
+
+interface Props {
+  t: TranslateType;
+  audio: Audio;
+  speech: boolean;
+}
+
+const AudioPlayerMounter: React.FC<Props> = ({ t, audio, speech }) => {
+  useEffect(() => {
+    initAudioPlayers();
+  }, []);
+
+  const locale = 'nb';
+  const license = getLicenseByAbbreviation(
+    audio.copyright.license.license,
+    locale,
+  );
+  const figureLicenseDialogId = `edit-audio-${audio.id}`;
+
+  const messages = {
+    title: t('dialog.title'),
+    close: t('dialog.close'),
+    rulesForUse: t('dialog.audio.rulesForUse'),
+    learnAboutLicenses: t('dialog.learnAboutLicenses'),
+    source: t('dialog.source'),
+  };
+
+  return (
+    <div>
+      <AudioPlayer
+        type={audio.audioFile.mimeType}
+        src={audio.audioFile.url}
+        title={audio.title}
+        speech={speech}
+      />
+      {!speech && (
+        <Fragment>
+          <FigureCaption
+            id={figureLicenseDialogId}
+            figureId={`figure-${audio.id}`}
+            caption={audio.caption}
+            reuseLabel=""
+            licenseRights={license.rights}
+            authors={audio.copyright.creators}
+          />
+          <FigureLicenseDialog
+            id={figureLicenseDialogId}
+            title={audio.title}
+            license={license}
+            authors={[]}
+            origin={origin}
+            messages={messages}
+            locale={locale}
+          />
+        </Fragment>
+      )}
+    </div>
+  );
+};
+
+export default injectT(AudioPlayerMounter);
